requirements.txt
flask==3.1.2
flask-cors==6.0.1
deep-translator==1.11.4
openai==1.101.0
langdetect==1.0.9
pymongo==4.14.1
requests==2.32.5
gunicorn==23.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
